fix(perf): reduce number of queries when loading an application page (#3239)
Refactor application authorization mechanism at Servlet level. Remove a search query in authorization mechanism.
If for some reason, no page is found for a given page token, a 404 is displayed which seems reasonable (instead of a 403).
I also replaced a search query with a get query to remove the additional count query performed by a search.

Cache application by token requests to reduce the database access. We can't rely on L2 cache for this kind of query so I used the `CacheService` as a cache given `applicationToken` are unique across the system.

Closes [RUNTIME-1885](https://bonitasoft.atlassian.net/browse/RUNTIME-1885)